Moved staff commands to text folder
