+++ src/main/java/com/honiism/discord/lemi/commands/text/staff/admins/ResetCurrData.java
@@ -17,22 +17,20 @@
  * along with Lemi-Bot. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.honiism.discord.lemi.commands.slash.staff.admins;
+package com.honiism.discord.lemi.commands.text.staff.admins;
 
 import java.util.HashMap;
 
-import com.honiism.discord.lemi.Config;
-import com.honiism.discord.lemi.Lemi;
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.honiism.discord.lemi.commands.handler.CommandCategory;
 import com.honiism.discord.lemi.commands.handler.UserCategory;
 import com.honiism.discord.lemi.commands.slash.handler.SlashCmd;
+import com.honiism.discord.lemi.data.UserDataManager;
 import com.honiism.discord.lemi.utils.misc.Tools;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.Permission;
+import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.User;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import net.dv8tion.jda.api.interactions.InteractionHook;
@@ -40,18 +38,17 @@ import net.dv8tion.jda.api.interactions.commands.OptionMapping;
 import net.dv8tion.jda.api.interactions.commands.OptionType;
 import net.dv8tion.jda.api.interactions.commands.build.Commands;
 
-public class ShardRestart extends SlashCmd {
+public class ResetCurrData extends SlashCmd {
 
-    private static final Logger log = LoggerFactory.getLogger(ShardRestart.class);
     private HashMap<Long, Long> delay = new HashMap<>();
     private long timeDelayed;
 
-    public ShardRestart() {
-        setCommandData(Commands.slash("shardrestart", "Restart a shard if it gets stuck.")
-                .addOption(OptionType.INTEGER, "shard_id", "The shard id to restart", false)
+    public ResetCurrData() {
+        setCommandData(Commands.slash("resetcurrdata", "Reset a user's currency data.")
+                .addOption(OptionType.USER, "user", "The user you want to reset.", true)
         );
 
-        setUsage("/admins shardrestart [shard id]");
+        setUsage("/admins resetcurrdata <user>");
         setCategory(CommandCategory.ADMINS);
         setUserCategory(UserCategory.ADMINS);
         setUserPerms(new Permission[] {Permission.ADMINISTRATOR});
@@ -60,53 +57,37 @@ public class ShardRestart extends SlashCmd {
     }
 
     @Override
-    public void action(SlashCommandInteractionEvent event) {
+    public void action(SlashCommandInteractionEvent event) throws JsonProcessingException {
         InteractionHook hook = event.getHook();
         User author = event.getUser();
-
+        
         if (delay.containsKey(author.getIdLong())) {
             timeDelayed = System.currentTimeMillis() - delay.get(author.getIdLong());
         } else {
             timeDelayed = (10 * 1000);
         }
             
-        if (timeDelayed >= (10 * 1000)) {        
+        if (timeDelayed >= (10 * 1000)) {
             if (delay.containsKey(author.getIdLong())) {
                 delay.remove(author.getIdLong());
             }
         
             delay.put(author.getIdLong(), System.currentTimeMillis());
 
-            Integer shardId = event.getOption("shard_id", OptionMapping::getAsInt);
+            Member member = event.getOption("user", OptionMapping::getAsMember);
 
-            if (shardId == null) {
-                log.info(author.getIdLong() + " is restarting all the shards.");
-
-                hook.sendMessage(":tulip: Restarting all the shards, see you in a bit :).").queue();
-
-                Lemi.getInstance().getShardManager().getGuildById(Config.get("honeys_hive"))
-                    .getTextChannelById(Config.get("logs_channel_id"))
-                    .sendMessage(author.getAsMention() + "Someone has reset all the shards. (<@" + author.getIdLong() + ">)")
-                    .queue(
-                        (success) -> {
-                            Lemi.getInstance().getShardManager().restart();
-                        }
-                    );
-                        
-            } else if (shardId != null && shardId < Lemi.getInstance().getShardManager().getShardsTotal()) {
-                log.info(author.getIdLong() + " is restarting the shard(" + shardId +").");
-                
-                hook.sendMessage(":tulip: Restarting the shard(" + shardId + "), see you in a bit :).").queue();
-
-                Lemi.getInstance().getShardManager().getGuildById(Config.get("honeys_hive"))
-                    .getTextChannelById(Config.get("logs_channel_id"))
-                    .sendMessage(author.getAsMention() + " is restarting the shard(" + shardId +").")
-                    .queue(
-                        (success) -> {
-                            Lemi.getInstance().getShardManager().restart(shardId);
-                        }
-                    );
+            if (member == null) {
+                hook.sendMessage(":grapes: That user doesn't exist in the guild.").queue();
+                return;
             }
+
+            setUserDataManager(member.getIdLong());
+
+            UserDataManager dataManager = getUserDataManager();
+
+            dataManager.removeData();
+
+            hook.sendMessage(":tulip: Successfully reset the user's data.").queue();
 
         } else {
             String time = Tools.secondsToTime(((10 * 1000) - timeDelayed) / 1000);
@@ -119,6 +100,6 @@ public class ShardRestart extends SlashCmd {
                 .setColor(0xffd1dc);
                 
             hook.sendMessageEmbeds(cooldownMsgEmbed.build()).queue();
-        }
-    }
+        }       
+    }    
 }
