+++ src/main/java/com/honiism/discord/lemi/commands/text/staff/dev/ModifyAdmins.java
@@ -17,7 +17,7 @@
  * along with Lemi-Bot. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.honiism.discord.lemi.commands.slash.staff.admins;
+package com.honiism.discord.lemi.commands.text.staff.dev;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -35,6 +35,7 @@ import com.honiism.discord.lemi.utils.paginator.Paginator;
 
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.Permission;
+import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.User;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
@@ -44,28 +45,28 @@ import net.dv8tion.jda.api.interactions.commands.OptionType;
 import net.dv8tion.jda.api.interactions.commands.build.Commands;
 import net.dv8tion.jda.api.interactions.commands.build.SubcommandData;
 
-public class UserBan extends SlashCmd {
+public class ModifyAdmins extends SlashCmd {
 
     private HashMap<Long, Long> delay = new HashMap<>();
     private long timeDelayed;
 
-    public UserBan() {
-        setCommandData(Commands.slash("userban", "Bans a user from using Lemi bot.")
+    public ModifyAdmins() {
+        setCommandData(Commands.slash("modifyadmins", "Add/remove/view user(s) to/from the administrator database.")
                 .addSubcommands(
-                        new SubcommandData("add", "Ban a user from using Lemi.")
-                                .addOption(OptionType.USER, "user", "The user you want to ban.", true)
-                                .addOption(OptionType.STRING, "reason", "The reason why they're getting banned.", true),
+                        new SubcommandData("add", "Add a user's key in the official administrator list.")
+                                .addOption(OptionType.USER, "user", "The user you want to add.", true)
+                                .addOption(OptionType.STRING, "key", "The key that will be assigned for this user.", true),
+  
+                        new SubcommandData("remove", "Remove a user from the official administrator list.")
+                                .addOption(OptionType.USER, "user", "The user you want to remove.", true),
 
-                        new SubcommandData("remove", "Unban a previously banned user.")
-                                .addOption(OptionType.USER, "user", "The user you want to unban.", true),
-
-                        new SubcommandData("view", "View all details from the ban list.")
+                        new SubcommandData("view", "View all details from the official administrator list.")
                 )
         );
 
-        setUsage("/userban ((subcommands))");
-        setCategory(CommandCategory.ADMINS);
-        setUserCategory(UserCategory.ADMINS);
+        setUsage("/dev modifyadmins ((subcommands))");
+        setCategory(CommandCategory.DEV);
+        setUserCategory(UserCategory.DEV);
         setUserPerms(new Permission[] {Permission.ADMINISTRATOR});
         setBotPerms(new Permission[] {Permission.ADMINISTRATOR});
         
@@ -75,7 +76,8 @@ public class UserBan extends SlashCmd {
     public void action(SlashCommandInteractionEvent event) {
         InteractionHook hook = event.getHook();
         User author = event.getUser();
-
+        Guild guild = event.getGuild();
+        
         if (delay.containsKey(author.getIdLong())) {
             timeDelayed = System.currentTimeMillis() - delay.get(author.getIdLong());
         } else {
@@ -94,14 +96,14 @@ public class UserBan extends SlashCmd {
             switch (subCmdName) {
                 case "add":
                     Member targetMember = event.getOption("user", OptionMapping::getAsMember);
-                    String reason = event.getOption("reason", OptionMapping::getAsString);
-                    
+                    String adminKey = event.getOption("key", OptionMapping::getAsString);
+
                     if (targetMember == null) {
                         hook.sendMessage(":grapes: That user doesn't exist in the guild.").queue();
                         return;
                     }
 
-                    LemiDbManager.INS.addBannedUserId(targetMember, reason, event);
+                    LemiDbManager.INS.addAdminId(guild, targetMember, adminKey, event);
                     break;
 
                 case "remove":
@@ -112,49 +114,49 @@ public class UserBan extends SlashCmd {
                         return;
                     }
 
-                    LemiDbManager.INS.removeBannedUserId(targetMember, event);
+                    LemiDbManager.INS.removeAdminId(guild, targetMember, event);
                     break;
 
                 case "view":
-                    viewAllBans(event);
+                    viewAllIds(event);
+                    break;
             }
+
         } else {
             String time = Tools.secondsToTime(((10 * 1000) - timeDelayed) / 1000);
                 
             EmbedBuilder cooldownMsgEmbed = new EmbedBuilder()
                 .setDescription("‧₊੭ :cherries: CHILL! ♡ ⋆｡˚\r\n" 
                         + "˚⊹ ˚︶︶꒷︶꒷꒦︶︶꒷꒦︶ ₊˚⊹.\r\n"
-                        + author.getAsMention() 
-                        + ", you can use this command again in `" + time + "`.")
+                        + author.getAsMention() + ", you can use this command again in `" + time + "`.")
                 .setColor(0xffd1dc);
                 
             hook.sendMessageEmbeds(cooldownMsgEmbed.build()).queue();
         }
     }
 
-    private void viewAllBans(SlashCommandInteractionEvent event) {
+    private void viewAllIds(SlashCommandInteractionEvent event) {
         InteractionHook hook = event.getHook();
-        List<String> banDetails = new ArrayList<>();
-        List<Long> authorIds = LemiDbManager.INS.getBannerAuthorIds(event);
-        List<Long> bannedUserIds = LemiDbManager.INS.getBannedUserIds(event);
-        List<String> reasons = LemiDbManager.INS.getBannedReasons(event);
+        List<String> adminDetails = new ArrayList<>();
+        List<Long> adminIds = LemiDbManager.INS.getAdminIds();
+        List<String> adminKeys = LemiDbManager.INS.getAdminKeys();
 
-        for (int i = 0; i < bannedUserIds.size(); i++) {
-            banDetails.add("Admin : <@" + authorIds.get(i) + ">" 
-                    + " | Banned user : <@" + bannedUserIds.get(i) + ">"
-                    + " | Reason : `" + reasons.get(i) + "`");
+        for (int i = 0; i < adminIds.size(); i++) {
+            adminDetails.add("<@" + adminIds.get(i) + "> `" 
+                    + adminIds.get(i) + " | key :` ||" 
+                    + adminKeys.get(i) + "||");
         }
 
-        if (Tools.isEmpty(banDetails)) {
-            hook.editOriginal(":fish_cake: There's no banned users.").queue();
+        if (Tools.isEmpty(adminDetails)) {
+            hook.editOriginal(":tulip: There's no admins.").queue();
             return;
         }
 
         Paginator.Builder builder = new Paginator.Builder(event.getJDA())
-            .setEmbedDesc("‧₊੭ :bread: **BANNED LIST!** ♡ ⋆｡˚")
             .setEventWaiter(Lemi.getInstance().getEventWaiter())
+            .setEmbedDesc("‧₊੭ :cake: **ADMINS!** ♡ ⋆｡˚")
             .setItemsPerPage(10)
-            .setItems(banDetails)
+            .setItems(adminDetails)
             .useNumberedItems(true)
             .useTimestamp(true)
             .addAllowedUsers(event.getUser().getIdLong())
@@ -163,7 +165,11 @@ public class UserBan extends SlashCmd {
 
         int page = 1;
 
-        hook.sendMessageEmbeds(EmbedUtils.getSimpleEmbed(":tea: Loading..."))
-            .queue(message -> builder.build().paginate(message, page));
+        event.getUser().openPrivateChannel().queue((privChannel) -> {
+            privChannel.sendMessageEmbeds(EmbedUtils.getSimpleEmbed(":seedling: Loading..."))
+                .queue(message -> builder.build().paginate(message, page));
+        });
+
+        hook.editOriginal(":snowflake: Sent you the details.").queue();
     }
 }
