+++ src/main/java/com/honiism/discord/lemi/commands/text/staff/dev/SetDebug.java
@@ -1,0 +1,105 @@
+/**
+ * Copyright (C) 2022 Honiism
+ * 
+ * This file is part of Lemi-Bot.
+ * 
+ * Lemi-Bot is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * Lemi-Bot is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with Lemi-Bot. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package com.honiism.discord.lemi.commands.text.staff.dev;
+
+import java.util.HashMap;
+
+import com.honiism.discord.lemi.Config;
+import com.honiism.discord.lemi.Lemi;
+import com.honiism.discord.lemi.commands.handler.CommandCategory;
+import com.honiism.discord.lemi.commands.handler.UserCategory;
+import com.honiism.discord.lemi.commands.slash.handler.SlashCmd;
+import com.honiism.discord.lemi.utils.misc.Tools;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import net.dv8tion.jda.api.EmbedBuilder;
+import net.dv8tion.jda.api.Permission;
+import net.dv8tion.jda.api.entities.User;
+import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
+import net.dv8tion.jda.api.interactions.InteractionHook;
+import net.dv8tion.jda.api.interactions.commands.OptionMapping;
+import net.dv8tion.jda.api.interactions.commands.OptionType;
+import net.dv8tion.jda.api.interactions.commands.build.Commands;
+
+public class SetDebug extends SlashCmd {
+
+    private static final Logger log = LoggerFactory.getLogger(SetDebug.class);
+    private HashMap<Long, Long> delay = new HashMap<>();
+    private long timeDelayed;
+
+    public SetDebug() {
+        setCommandData(Commands.slash("setdebug", "Toggle debug mode for Lemi.")
+                .addOption(OptionType.BOOLEAN, "value", "Set the debug mode, true = on | false = off", true)
+        );
+
+        setUsage("/dev compile <true|false>");
+        setCategory(CommandCategory.DEV);
+        setUserCategory(UserCategory.DEV);
+        setUserPerms(new Permission[] {Permission.ADMINISTRATOR});
+        setBotPerms(new Permission[] {Permission.ADMINISTRATOR});
+    }
+
+    @Override
+    public void action(SlashCommandInteractionEvent event) {
+        InteractionHook hook = event.getHook();
+        User author = event.getUser();
+
+        if (delay.containsKey(author.getIdLong())) {
+            timeDelayed = System.currentTimeMillis() - delay.get(author.getIdLong());
+        } else {
+            timeDelayed = (10 * 1000);
+        }
+            
+        if (timeDelayed >= (10 * 1000)) {
+            if (delay.containsKey(author.getIdLong())) {
+                delay.remove(author.getIdLong());
+            }
+        
+            delay.put(author.getIdLong(), System.currentTimeMillis());
+
+            boolean debugValue = event.getOption("value", OptionMapping::getAsBoolean);
+
+            Lemi.getInstance().setDebug(debugValue);
+
+            String logMsg = "Successfully toggled debug mode to " + (debugValue ? "on" : "off") + "!";
+
+            hook.sendMessage(":tulip: " + logMsg).queue();
+            log.info(logMsg);
+
+            Lemi.getInstance().getShardManager().getGuildById(Config.get("honeys_hive"))
+        	.getTextChannelById(Config.get("logs_channel_id"))
+        	.sendMessage(logMsg)
+                .queue();
+        } else {
+            String time = Tools.secondsToTime(((5 * 1000) - timeDelayed) / 1000);
+                
+            EmbedBuilder cooldownMsgEmbed = new EmbedBuilder()
+                .setDescription("‧₊੭ :cherries: CHILL! ♡ ⋆｡˚\r\n" 
+                        + "˚⊹ ˚︶︶꒷︶꒷꒦︶︶꒷꒦︶ ₊˚⊹.\r\n"
+                        + author.getAsMention() 
+                        + ", you can use this command again in `" + time + "`.")
+                .setColor(0xffd1dc);
+                
+            hook.sendMessageEmbeds(cooldownMsgEmbed.build()).queue();
+        }        
+    }    
+}
