+++ src/main/java/com/honiism/discord/lemi/commands/text/staff/mods/Dashboard.java
@@ -17,7 +17,7 @@
  * along with Lemi-Bot. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.honiism.discord.lemi.commands.slash.staff.mods;
+package com.honiism.discord.lemi.commands.text.staff.mods;
 
 import java.lang.management.ManagementFactory;
 import java.lang.management.RuntimeMXBean;
