+++ src/main/java/com/honiism/discord/lemi/commands/text/staff/mods/ShardStatus.java
@@ -10,16 +10,17 @@
  * 
  * Lemi-Bot is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  * 
  * You should have received a copy of the GNU General Public License
- * along with Lemi-Bot.  If not, see <http://www.gnu.org/licenses/>.
+ * along with Lemi-Bot. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.honiism.discord.lemi.commands.slash.staff.mods;
+package com.honiism.discord.lemi.commands.text.staff.mods;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
@@ -28,12 +29,12 @@ import com.honiism.discord.lemi.Lemi;
 import com.honiism.discord.lemi.commands.handler.CommandCategory;
 import com.honiism.discord.lemi.commands.handler.UserCategory;
 import com.honiism.discord.lemi.commands.slash.handler.SlashCmd;
-import com.honiism.discord.lemi.data.items.Items;
 import com.honiism.discord.lemi.utils.misc.EmbedUtils;
 import com.honiism.discord.lemi.utils.misc.Tools;
 import com.honiism.discord.lemi.utils.paginator.Paginator;
 
 import net.dv8tion.jda.api.EmbedBuilder;
+import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.User;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
@@ -42,17 +43,17 @@ import net.dv8tion.jda.api.interactions.commands.OptionMapping;
 import net.dv8tion.jda.api.interactions.commands.OptionType;
 import net.dv8tion.jda.api.interactions.commands.build.Commands;
 
-public class ViewItems extends SlashCmd {
+public class ShardStatus extends SlashCmd {
 
     private HashMap<Long, Long> delay = new HashMap<>();
     private long timeDelayed;
 
-    public ViewItems() {
-        setCommandData(Commands.slash("viewitems", "View the currently available items in the internal list.")
-                .addOption(OptionType.INTEGER, "page", "The page number for the items list you want to see.", false)
+    public ShardStatus() {
+        setCommandData(Commands.slash("shardstatus", "View the status of all shards.")
+                .addOption(OptionType.INTEGER, "page", "The page number for the shard status you want to see.", false)
         );
 
-        setUsage("/mods viewitems [page number]");
+        setUsage("/mods shardstatus [page number]");
         setCategory(CommandCategory.MODS);
         setUserCategory(UserCategory.MODS);
         setUserPerms(new Permission[] {Permission.MESSAGE_MANAGE});
@@ -68,44 +69,52 @@ public class ViewItems extends SlashCmd {
         if (delay.containsKey(author.getIdLong())) {
             timeDelayed = System.currentTimeMillis() - delay.get(author.getIdLong());
         } else {
-            timeDelayed = (5 * 1000);
+            timeDelayed = (10 * 1000);
         }
             
-        if (timeDelayed >= (5 * 1000)) {
+        if (timeDelayed >= (10 * 1000)) {
             if (delay.containsKey(author.getIdLong())) {
                 delay.remove(author.getIdLong());
             }
         
             delay.put(author.getIdLong(), System.currentTimeMillis());
-            
-            List<String> items = new ArrayList<String>();
 
-            for (Items item : Items.getItems().values()) {
-                if (Items.checkIfItemExists(item.getName())) {
-                   items.add(item.getEmoji() + " " + item.getName() + " | " + item.getId()); 
-                } else {
-                    items.add(item.getEmoji() + " " + item.getName() + " | " + item.getId() + " | **NOT IN LIST**");
-                }
+            List<JDA> shards = new ArrayList<>(Lemi.getInstance().getShardManager().getShardCache().asList());
+            Collections.reverse(shards);
+
+            List<String> shardDetailsItems = new ArrayList<>();
+
+            for (JDA shard : shards) {
+                String shardDetails = "Shard id : " + shard.getShardInfo().getShardId()
+                        + " | status : " + shard.getStatus()
+                        + " | cached guilds : " + shard.getGuildCache().size()
+                        + " | cached members : " + shard.getUserCache().size()
+                        + " | gateway ping : " + shard.getGatewayPing() + " ms";
+
+                shardDetailsItems.add(shardDetails);
             }
 
             Paginator.Builder builder = new Paginator.Builder(event.getJDA())
+                .setEmbedDesc("‧₊੭ :candy: **SHARD STATUS** ♡ ⋆｡˚\r\n"
+                        + ":sunflower: Total shards : " + shards.size() + "\r\n"
+                        + ":seedling: Total guilds : " + Lemi.getInstance().getShardManager().getGuilds().size() + "\r\n"
+                        + ":snowflake: Cached users : " + Lemi.getInstance().getShardManager().getUserCache().size() + "\r\n")
                 .setEventWaiter(Lemi.getInstance().getEventWaiter())
-                .setEmbedDesc("‧₊੭ :tulip: **ITEMS!** ♡ ⋆｡˚")
                 .setItemsPerPage(10)
-                .setItems(items)
+                .setItems(shardDetailsItems)
                 .useNumberedItems(true)
                 .useTimestamp(true)
-                .addAllowedUsers(author.getIdLong())
+                .addAllowedUsers(event.getUser().getIdLong())
                 .setColor(0xffd1dc)
                 .setTimeout(1, TimeUnit.MINUTES);
 
             int page = event.getOption("page", 1, OptionMapping::getAsInt);
 
-            hook.sendMessageEmbeds(EmbedUtils.getSimpleEmbed(":seedling: Loading..."))
+            hook.sendMessageEmbeds(EmbedUtils.getSimpleEmbed(":umbrella2: Loading..."))
                 .queue(message -> builder.build().paginate(message, page));
-                
+
         } else {
-            String time = Tools.secondsToTime(((5 * 1000) - timeDelayed) / 1000);
+            String time = Tools.secondsToTime(((10 * 1000) - timeDelayed) / 1000);
                 
             EmbedBuilder cooldownMsgEmbed = new EmbedBuilder()
                 .setDescription("‧₊੭ :cherries: CHILL! ♡ ⋆｡˚\r\n" 
@@ -115,6 +124,6 @@ public class ViewItems extends SlashCmd {
                 .setColor(0xffd1dc);
                 
             hook.sendMessageEmbeds(cooldownMsgEmbed.build()).queue();
-        }
-    }
+        }       
+    }    
 }
